Use refinements for core ext
