+++ src/ruby/yast/core_ext.rb
@@ -1,1 +1,1 @@
-require_relative "core_ext/string"
+require_relative "core_ext/ansi_string"
