--- src/ruby/yast/core_ext/string.rb
@@ -1,1 +1,0 @@
-require_relative "string/ansi"
